% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata.samplesize.R
\name{strata.samplesize}
\alias{strata.samplesize}
\title{Sample size estimation on stratified sampling}
\usage{
strata.samplesize(
  Nh,
  var,
  error,
  alpha,
  estimator = c("total", "mean", "proportion", "class total"),
  alloc = c("prop", "min", "optim"),
  ch,
  p,
  mean,
  replace = FALSE,
  relative = FALSE
)
}
\arguments{
\item{Nh}{Vector of population strata sizes.}

\item{var}{Vector of estimated strata variances.}

\item{error}{Sampling error.}

\item{alpha}{Significance level to obtain confidence intervals.}

\item{estimator}{The estimator to be estimated. Default is "total".}

\item{alloc}{The allocation to be used when taking samples. Default is "prop".}

\item{ch}{Vector of cost per strata to select an individual for the sample.}

\item{p}{Estimated population proportion. If estimator is not "proportion" or "class total" it will be ignored.}

\item{mean}{Estimated population mean. If relative=FALSE it will be ignored.}

\item{replace}{Whether the samples to be taken can have repeated instances or not.}

\item{relative}{Whether the specified error is relative or not.}
}
\value{
Number of instances of the sample to be taken.
}
\description{
Calculates the required sample size in order to achieve an
absolute or relative sampling error  less or equal to the specified for an specific
estimator and an optional confidence interval in stratified sampling.
}
\details{
With "proportion" and "class total" estimators variance vector must
contain \code{\link{var}} return values equal to \eqn{\frac{Nh}{(Nh-1)}p*(1-p)} values.
}
\examples{
strata.samplesize(c(120,100,110,50), c(458, 313,407,364), error=5, alpha=0.05, "mean", "prop")

}
