% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.utils.R
\name{sys.all.samples}
\alias{sys.all.samples}
\title{Systematic samples}
\usage{
sys.all.samples(data, n)
}
\arguments{
\item{data}{Population data}

\item{n}{Sample size}
}
\value{
List with a sample per entrance
}
\description{
Returns all possible systematic samples of size n
}
\examples{
data<-c(1,3,5,2,4,6,2,7,3)
sys.all.samples(data, 3)
}
