% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.utils.R
\name{sys.corrwst}
\alias{sys.corrwst}
\title{Stratified correlation coefficient}
\usage{
sys.corrwst(N, n, data)
}
\arguments{
\item{N}{Population size}

\item{n}{Sample size}

\item{data}{Population data}
}
\value{
Correlation coefficient
}
\description{
Stratified correlation coefficient
}
\examples{
data<-c(1,3,5,2,4,6,2,7,3)
sys.corrwst(9,3,data)  #0.09022556
}
