% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys.utils.R
\name{sys.intracorr}
\alias{sys.intracorr}
\title{Intraclass correlation coefficient}
\usage{
sys.intracorr(N, n, data)
}
\arguments{
\item{N}{Population size}

\item{n}{Sample size}

\item{data}{Population data}
}
\value{
Intraclass correlation
}
\description{
Intraclass correlation coefficient
}
\examples{
data<-c(1,3,5,2,4,6,2,7,3)
sys.intracorr(9, 3, data)  #0.34375 example 1
}
