% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syst.utils.R
\name{syst.anova}
\alias{syst.anova}
\title{Analysis of variance of population data}
\usage{
syst.anova(data, n)
}
\arguments{
\item{data}{Population data}

\item{n}{Sample size}
}
\value{
Summary
}
\description{
Analysis of variance of population data
}
\examples{
data<-c(1,3,5,2,4,6,2,7,3)
syst.anova(data,3)
}
