% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syst.sample.R
\name{syst.sample}
\alias{syst.sample}
\title{Systematic sampling sample}
\usage{
syst.sample(N, n, data)
}
\arguments{
\item{N}{Population size.}

\item{n}{Sample size}

\item{data}{Optional data of the population.}
}
\value{
Vector of size n with numbers from 1 to N indicating the index samples to be taken.
If data is provided then the instances will be returned.
}
\description{
Retrieves a \eqn{\frac{N}{n}} systematic sample
}
\details{
If \eqn{\frac{N}{n}} is not an even number a 1 in floor(\eqn{\frac{N}{n}}) sample will be taken.
}
\examples{
data<-runif(40)
syst.sample(40,8, data)

}
